% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNormalChromosomes.R
\name{computeNormalChromosomes}
\alias{computeNormalChromosomes}
\title{computeNormalChromosomes}
\usage{
computeNormalChromosomes(segments, perc = 0.15)
}
\arguments{
\item{segments}{data.frame formatted with correct column names}

\item{perc}{decimal value of alteration frequency, by default is 0.15}
}
\value{
vector with chormosome names
}
\description{
computeNormalChromosomes
}
\examples{
\dontrun{
computeNormalChromosomes(segments)
}
}
