% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCluster.R
\name{PlotCluster}
\alias{PlotCluster}
\title{PlotCluster}
\usage{
PlotCluster(segs, clust_method = "ward.D2", plot_path)
}
\arguments{
\item{segs}{data.frame with segments of samples. It must be formatted with correct column names (start, end, ID)}

\item{clust_method}{clustering method. Default is "ward.D2"}

\item{plot_path}{Path to save output plots}
}
\value{
Plot with chromosomes clustered
}
\description{
The function clusters chromosomes based on the copy number (CN) and returns a graph where it is possible to observe the different groups
}
