% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNormalChromosomes.R
\name{computeNormalChromosomes}
\alias{computeNormalChromosomes}
\title{computeNormalChromosomes}
\usage{
computeNormalChromosomes(
  segments,
  perc = 0.15,
  maxCN = 6,
  min_threshold = 1.6,
  max_threshold = 2.4
)
}
\arguments{
\item{segments}{data.frame formatted with correct column names}

\item{perc}{decimal value of alteration frequency, by default is 0.15}

\item{maxCN}{threshold of max copy number to consider, by default is 6}

\item{min_threshold}{minimum threshold to define a normal CN, by default is 1.60}

\item{max_threshold}{maximum threshold to define a normal CN, by default is 2.40}
}
\value{
vector with chormosome names
}
\description{
computeNormalChromosomes
}
\examples{
\dontrun{
computeNormalChromosomes(segments)
}
}
