% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRrefit.R
\name{DRrefit}
\alias{DRrefit}
\title{DRrefit}
\usage{
DRrefit(
  segments_chort,
  chrlist,
  maxCN = 6,
  clust_method = "ward.D2",
  plot_output = FALSE,
  plot_path
)
}
\arguments{
\item{segments_chort}{data.frame formatted with correct column names}

\item{chrlist}{list of normal chromosome arms (pathology-specific)}

\item{maxCN}{threshold of max copy number to consider. By default is 6}

\item{clust_method}{clustering method. By default is "ward.D2"}

\item{plot_output}{Whether to plot refitted profiles (logical)}

\item{plot_path}{Path to save output plots}
}
\value{
A plot that shows old and refitted copy number profiles  and two databases (segment corrected and report)
}
\description{
This function refits the diploid region of input copy number profiles (BED file)
}
