% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRrefit_plot.R
\name{DRrefit_plot}
\alias{DRrefit_plot}
\title{DRrefit_plot}
\usage{
DRrefit_plot(segments_refitted, DRrefit_report, plot_format = "png", plot_path)
}
\arguments{
\item{segments_refitted}{DRrefit output dataframe}

\item{DRrefit_report}{DRrefit output dataframe}

\item{plot_format}{file format for the output plots. By default is "png" (accepts "png", "jpg", "pdf", "tif")}

\item{plot_path}{path to save output plots}
}
\value{

}
\description{
DRrefit_plot
}
\examples{
#' chr_list <- c("10q","11p","12p","19q","1p","21q","2q","3p","4p","4q","6p","6q","7p" )
results <- DRrefit(segments,chrlist = chr_list)
DRrefit_report <- results$report
segments_refitted <- results$segments_corrected
DRrefit_plot(segments_refitted,DRrefit_report,plot_path)


}
