% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Popeye.R
\name{Popeye}
\alias{Popeye}
\title{Popeye}
\usage{
Popeye(segments)
}
\arguments{
\item{segments}{data.frame formatted with correct column names (see package vignette)}
}
\value{
Return a data frame containg segments with the arm annotation.
}
\description{
The function assign the chromosomal arm to each segment.
}
\examples{
data <- Popeye(segments)
}
